First pass at the Lifelogging Camera
It's capturing the streams as expected. I've been having trouble trying to get ffmpeg installed on the thing, but if I use it locally on my mac it encodes the outputs quickly and just fine. Still working on encode_recordings.py so that it takes the file name in properly before making the CLI call.